Refactor: move deploy helpers into dedicated lib
Related: DEVOPS-119
